Use DynamoDb test fixture (#39)
