+++ ActivityListener.Tests/AppTestCollection.cs
@@ -1,0 +1,18 @@
+using Amazon.DynamoDBv2;
+using Amazon.DynamoDBv2.DataModel;
+using Amazon.DynamoDBv2.Model;
+using Microsoft.Extensions.Hosting;
+using System;
+using System.Collections.Generic;
+using Xunit;
+
+namespace ActivityListener.Tests
+{
+    [CollectionDefinition("AppTest collection", DisableParallelization = true)]
+    public class AppTestCollection : ICollectionFixture<MockApplicationFactory>
+    {
+        // This class has no code, and is never created. Its purpose is simply
+        // to be the place to apply [CollectionDefinition] and all the
+        // ICollectionFixture<> interfaces.
+    }
+}
